Empty $this->aliases lead to error in Container when using Component standalone
When using this component standalone and dumping the container, the aliases are not set to the default value of $this->aliases = array(). If now calling $container->has() it is checked if $this->alias has a key: array_key_exists($id, $this->aliases). array_key_exists needs an array as second parameter what won't be set if no aliases are set during the container build process. So instead of returning an empty string, returning the initialization value helps to avoid this error.